+++ FinalTest.java
@@ -1,0 +1,13 @@
+public class FinalTest
+{
+    public void test()
+    {
+        for(i = 0;i < size;i + +)
+        {
+            if(value = = 1)
+            {
+                count + + ;
+            }
+        }
+    }
+}