+++ SimpleTest.java
@@ -1,0 +1,12 @@
+public class SimpleTest
+{
+    public void test()
+    {
+        int i = 0;
+        i ++;
+        if(i == 1)
+        {
+            System.out.println("test");
+        }
+    }
+}