+++ TestBrokenOps.java
@@ -1,0 +1,13 @@
+public class TestBrokenOps
+{
+    public void test()
+    {
+        for(i = 0;i < super.zone.players.size();i + +)
+        {
+            if(this.aChar.size() = = 1)
+            {
+                break;
+            }
+        }
+    }
+}