+++ TestOperators.java
@@ -1,0 +1,24 @@
+public class TestOperators
+{
+    public void testMethod()
+    {
+        for(i = 0;i < super.zone.players.size();i + +)
+        {
+            if(this.aChar.size() = = 1)
+            {
+                break;
+            }
+            count + + ;
+            value - - ;
+            if(a = = b && c!= d)
+            {
+                System.out.println("Test");
+            }
+            // Test more operators
+            x + = 5;
+            y - = 3;
+            z * = 2;
+            w / = 4;
+        }
+    }
+}